Bump version in setup.py and modify handling of databases for Async
This commit both updates the version of the tool in setup.py and significantly modifies handling of database types in _database_obj.py. The update was necessitated to include handling for if a FeatureLocation doesn't exist and for hidden qualifiers in _parsing.py.

Adaptations were also made to support asynchronous operations in _orm.py, while _database_obj.py was refactored to support both regular and asynchronous database operations. Threading conditions, connection pooling and session handling were developed for async compatibility.

This will result in better performance when making concurrent requests or when the tool has to wait for IO-bound tasks.