Bump version to 2.1.0 and add async handler in _database_obj.py
Version of the application has been increased as part of the enhancements done on FeatureLocation and hidden qualifiers in _parsing.py.
Experimented with async context manager to handle sessions in _database_obj.py. This allows using `with` to manage session lifecycle. It led to code simplification as the session lifecycle management moved to Python`s context management.