Update version to 2.3.1 and optimize doc description retrieval
The GeneiousDB version in setup.py has been updated to 2.3.1. Also, the document description fetching process in '_database_obj.py' has been simplified to improve efficiency and maintenance. Now, instead of traversing through 'XMLSerialisableRootElement', the system fetches the description directly from the document. If a description is not found, it defaults to the document name